+++ RELEASE_NOTES.md
@@ -1,1 +1,2 @@
+* Renderer: fix broken script link in plain language summaries
 * QA: Fix broken link checking